Seth, remainder of tabs recompartmentalized and functional.
